fix(custom-widget): reuse CE markdown renderer barrel
