--- app/client/src/components/editorComponents/GPT/shared/index.ts
@@ -1,1 +1,0 @@
-export { AIMarkdownRenderer } from "./AIMarkdownRenderer";
